• More things for the complete GDS data model, for finally supporting CBOR :> 	• Like ValueType! :D

+++ src/rebound/math/MathUtilities.java
@@ -2062,7 +2062,7 @@ implements JavaNamespace
 		
 		
 		if (a instanceof Float)
-			return convertFloatToRationalOrInteger((Double)a);
+			return convertFloatToRationalOrInteger((Float)a);
 		if (a instanceof Double)
 			return convertFloatToRationalOrInteger((Double)a);
